Working project
Still need to confirm that the five minute timer works, but I'm fairly confident that it does. Next steps is to make it all a little prettier and add some custom graphics. After that it will be done until someone has some feature requests, such as showing some of the other data I grab like humidity levels and wind and precipitation.